Renamed src/ to lib/
