+++ lib/classes/Num2Word.mjs
