+++ lib/i18n/FR.mjs
