+++ lib/i18n/NO.mjs
